backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pymongo==3.12.0
python-dotenv==0.19.0
bcrypt==3.2.0
PyJWT==2.1.0
requests==2.26.0
python-magic==0.4.24
sendgrid==6.9.0
werkzeug==2.0.1 